+++ src/config/GeneralConfig.php
@@ -1669,21 +1669,12 @@ class GeneralConfig extends BaseObject
     public $verifyEmailSuccessPath = '';
 
     /**
-     * @var array Stores any custom config settings
-     */
-    private array $_customSettings = [];
-
-    /**
      * @inheritdoc
      */
     public function __get($name)
     {
         if (isset(self::$renamedSettings[$name])) {
             return $this->{self::$renamedSettings[$name]};
-        }
-
-        if (array_key_exists($name, $this->_customSettings)) {
-            return $this->_customSettings[$name];
         }
 
         return parent::__get($name);
@@ -1702,11 +1693,7 @@ class GeneralConfig extends BaseObject
             return;
         }
 
-        try {
-            parent::__set($name, $value);
-        } catch (UnknownPropertyException $e) {
-            $this->_customSettings[$name] = $value;
-        }
+        parent::__set($name, $value);
     }
 
     /**
@@ -1716,10 +1703,6 @@ class GeneralConfig extends BaseObject
     {
         if (isset(self::$renamedSettings[$name])) {
             return isset($this->{self::$renamedSettings[$name]});
-        }
-
-        if (array_key_exists($name, $this->_customSettings)) {
-            return $this->_customSettings[$name] !== null;
         }
 
         return parent::__isset($name);